requirements.txt
flask>=3.0.0
scikit-learn>=1.4.0
pdfminer.six>=20221105
nltk>=3.8.1
transformers>=4.35.2
torch>=2.1.1
numpy>=1.26.0

